refactor: tidy up server params and split reusable into abstract
